// src/types/interface.ts
/**
 * 坐标点
 */
export interface PointType {
  x: number
  y: number
}

/**
 * 编辑区域信息
 */
export interface AreaInfoType {
  width: number
  height: number
  isShowArea: boolean
}

/**
 * 样式
 */
export interface StyleType {
  left: number
  top: number
  width: number
  height: number
  rotate: number
}